+++ Java.EE/lesson2/src/main/java/gb/lesson2/servlet/MainServlet.java
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class Main extends HttpServlet {
+public class MainServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -14,13 +14,14 @@ public class Main extends HttpServlet {
         resp.getWriter().print("<h1>Mains page servlet</h1>");
         resp.getWriter().print("<table>");
         resp.getWriter().print("<tr>");
-        resp.getWriter().print("<td><a href=\"/lesson2/main\">Main</a></td>");
-        resp.getWriter().print("<td><a href=\"/lesson2/catalog\">Catalog</a></td>");
-        resp.getWriter().print("<td><a href=\"/lesson2/product\">Product</a></td>");
-        resp.getWriter().print("<td><a href=\"/lesson2/cart\">Cart</a></td>");
-        resp.getWriter().print("<td><a href=\"/lesson2/order\">Order</a></td>");
+        resp.getWriter().print("<td><a href=\"" + req.getContextPath() + "/main\">Main</a></td>");
+        resp.getWriter().print("<td><a href=\"" + req.getContextPath() + "/catalog\">Catalog</a></td>");
+        resp.getWriter().print("<td><a href=\"" + req.getContextPath() + "/product\">Product</a></td>");
+        resp.getWriter().print("<td><a href=\"" + req.getContextPath() + "/cart\">Cart</a></td>");
+        resp.getWriter().print("<td><a href=\"" + req.getContextPath() + "/order\">Order</a></td>");
         resp.getWriter().print("</tr>");
         resp.getWriter().print("</table>");
+
 
 
     }
